+++ native/src/seal/util/common.h
@@ -567,6 +567,6 @@ namespace seal
             return value == T{ 0 };
         }
 
-        void seal_memzero(void *const data, std::size_t size);
+        void seal_memzero(void *data, std::size_t size);
     } // namespace util
 } // namespace seal